refactor: remove unused folders, fix styled components
