+++ src/api/index.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { API_URL, TOKEN } from '../app/constants/authorization';
+import { API_URL, TOKEN } from '../constants/authorization';
 
 const api = axios.create({
   withCredentials: true,
