+++ src/components/Footer.tsx
@@ -1,20 +1,19 @@
-/* eslint-disable react/jsx-props-no-spreading */
 import React from 'react';
 import { AppBar, Grid } from '@mui/material';
-import team from '../../app/constants/teammate';
-import RSSchcoolLogo from '../RSSchcoolLogo/RSSchcoolLogo';
-import TeammateLink from '../TeammateLink/TeammateLink';
-import COPYRIGHT from '../../app/constants/copyright';
-import GridFlexGrow from '../GridFlexGrow/GridFlexGrow';
+import team from '../constants/teammate';
+import RSSchcoolLogo from '../assets/rs-school-js.svg';
+import TeammateLink from './TeammateLink';
+import COPYRIGHT from '../constants/copyright';
+import GridFlexGrow from './GridFlexGrow';
 
 function Footer(): JSX.Element {
   return (
-    <AppBar position="static" sx={{ padding: '15px' }}>
+    <AppBar position="static" sx={{ padding: '15px' }} component="footer">
       <Grid container spacing={2} alignItems="center" justifyContent="center">
         <Grid item>
           <RSSchcoolLogo />
         </Grid>
-        <GridFlexGrow>
+        <GridFlexGrow item>
           <Grid container spacing={2} justifyContent="center" alignItems="center">
             {team.map((item) => (
               <Grid item key={item.id}>
