+++ src/components/FormField.tsx
@@ -1,0 +1,8 @@
+import { TextField } from '@mui/material';
+import { styled } from '@mui/material/styles';
+
+const CustomizedTextField = styled(TextField)`
+  margin-bottom: 25px;
+`;
+
+export default CustomizedTextField;
