+++ src/components/GridFlexGrow.tsx
@@ -1,0 +1,8 @@
+import { Grid } from '@mui/material';
+import { styled } from '@mui/material/styles';
+
+const GridFlexGrow = styled(Grid)`
+  flex-grow: 1;
+`;
+
+export default GridFlexGrow;
