+++ src/components/Header.tsx
@@ -1,12 +1,12 @@
 import { AppBar, Grid } from '@mui/material';
 import React from 'react';
-import AppRoutes from '../../app/constants/routes';
-import { navText } from '../../app/constants/text';
-import { useAppSelector } from '../../hooks/reduxTypedHooks';
-import { authSelector } from '../../store/authSlice';
-import AuthButtonsContainer from '../AuthButtonsContainer/AuthButtonsContainer';
-import GridFlexGrow from '../GridFlexGrow/GridFlexGrow';
-import LinkStyled from '../LinkStyled/LinkStyled';
+import AppRoutes from '../constants/routes';
+import { navText } from '../constants/text';
+import { useAppSelector } from '../hooks/reduxTypedHooks';
+import { authSelector } from '../store/authSlice';
+import AuthButtonsContainer from './AuthButtonsContainer';
+import GridFlexGrow from './GridFlexGrow';
+import LinkStyled from './LinkStyled';
 
 const Header: React.FC = () => {
   const { userId } = useAppSelector(authSelector);
@@ -14,11 +14,11 @@ const Header: React.FC = () => {
   return (
     <AppBar position="static" sx={{ padding: '15px' }}>
       <Grid container spacing={2}>
-        <GridFlexGrow>
+        <GridFlexGrow item>
           <LinkStyled to={AppRoutes.WELCOME}>{navText.home}</LinkStyled>
         </GridFlexGrow>
         {userId && (
-          <GridFlexGrow>
+          <GridFlexGrow item>
             <LinkStyled to={AppRoutes.PROJECTS}>{navText.projects}</LinkStyled>
           </GridFlexGrow>
         )}
