+++ src/components/Layout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Footer from '../Footer/Footer';
-import Header from '../Header/Header';
+import Footer from './Footer';
+import Header from './Header';
 
 const Layout = () => (
   <>
