+++ src/constants/copyright.ts
