+++ src/constants/text.ts
